feat: Populate provided VPC with subnets so that a VPC can be looked up (#31)
Fixes #13 

Adds two constructs for populating a VPC provided by an enterprise.

Scenario is that the developer doesn't have permissions to create their own VPC, but are provided a VPC ID, and two route tables. One route table with routes to the enterprise network, and one route table for routing locally within the VPC.

`PopulateWithConfig` allows a user to pass an array of `SubnetConfig`s

`SplitVpcEvenly` allows a user to split the subnet evenly, using sensible defaults, but also providing options to customize