feat: include a pre-commit hook that adds API.md to the end of README.md (#67)
Fixes #44 

Some enterprises maintain their own mirrors of open source repos such as npm and pypi that developers within the enterprise must use.

Depending on how this is implemented, the source files from a project (including `API.md`) might not be available to developers, but the `README.md` _is_ available.

This PR adds a pre-commit hook to copy `API.md` to the end of `README.md` in a hidden section that a user can click to expand.

This will allow developers in enterprise settings that don't have access to github the ability to view the API documentation

The large diff to `README.md` is the appending of `API.md` to the end