fix: Resource Extractor allow RetentionPolicies (#294)
Fixes #293 

This pull request will allow Retention Policies to end up in the extracted stack for a given resource.

Use case: I want to extract KMS Keys to the extracted stack. I want to set a retention policy on these keys (RETAIN). Currently, this doesn't work. This pull request will fix this bug and allow these properties to correctly be set in the extracted stack resource.